insert SVG image and create Icon component
