--- src/shared/Button/Button.jsx
@@ -1,133 +1,0 @@
-import "../../css/button/button.css"
-
-export const Buttons = () => {
-  return (
-    <div className="buttons-wrapper">
-      <button className="button button_theme-blue button_size_medium">
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-        Text here
-      </button>
-      <button className="button button_theme-blue-transparent button_size_medium">
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-        Text here
-      </button>
-      <button className="button button_theme-black-transparent button_size_medium">
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-        Text here
-      </button>
-
-      <button className="button button_theme-blue button_size_medium">
-        Text here
-      </button>
-      <button className="button button_theme-blue-transparent button_size_medium">
-        Text here
-      </button>
-      <button className="button button_theme-black-transparent button_size_medium">
-        Text here
-      </button>
-
-      <button
-        className="button button_theme-blue button_size_medium button_icon-only"
-      >
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-      </button>
-      <button
-        className="button button_theme-blue-transparent button_size_medium button_icon-only"
-      >
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-      </button>
-      <button
-        className="button button_theme-black-transparent button_size_medium button_icon-only"
-      >
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-      </button>
-
-      <button className="button button_theme-blue button_size_small">
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-        Text here
-      </button>
-      <button className="button button_theme-blue-transparent button_size_small">
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-        Text here
-      </button>
-
-      <button className="button button_theme-blue button_size_small">
-        Text here
-      </button>
-      <button className="button button_theme-blue-transparent button_size_small">
-        Text here
-      </button>
-
-      <button
-        className="button button_theme-blue button_size_small button_icon-only"
-      >
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-      </button>
-      <button
-        className="button button_theme-blue-transparent button_size_small button_icon-only"
-      >
-        {/*<svg
-          className="icon"
-          viewBox="0 0 16 16"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M12.5 3.5L3.5 12.5M3.5 3.5L12.5 12.5" fill="none" />
-        </svg>*/}
-      </button>
-  </div>
-)
-}
